- Parse native.go into separate files for easier readability
